Move Data extraction to codegen
