+++ c2c-extension-test/src/test/resources/testng-integration.xml
@@ -31,6 +31,7 @@
         <classes>
             <class name="io.ballerina.c2c.test.JobTest"/>
             <class name="io.ballerina.c2c.test.EnvTest"/>
+            <class name="io.ballerina.c2c.test.ServiceExtractionTest"/>
             <class name="io.ballerina.c2c.test.CustomDiagnosticsTest"/>
             <class name="io.ballerina.c2c.test.codeactions.InvalidServicePath"/>
             <class name="io.ballerina.c2c.test.codeactions.InvalidResourcePath"/>
