+++ c2c-extension-test/src/test/resources/testng.xml
@@ -25,7 +25,7 @@
             </run>
         </groups>
         <classes>
-<!--            <class name="io.ballerina.c2c.test.samples.Sample1Test"/>-->
+            <class name="io.ballerina.c2c.test.samples.Sample1Test"/>
             <class name="io.ballerina.c2c.test.samples.Sample5Test"/>
             <class name="io.ballerina.c2c.test.samples.Sample6Test"/>
             <class name="io.ballerina.c2c.test.samples.Sample7Test"/>
